CNN/requirements.txt
args==0.1.0
asttokens==2.4.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
clint==0.5.1
colorama==0.4.6
comm==0.2.2
comtypes==1.1.14
debugpy==1.8.6
decorator==5.1.1
ecapture==2.0.2
exceptiongroup==1.2.2
executing==2.1.0
feedparser==6.0.10
Flask==2.2.3
idna==3.4
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.1.2
jaraco.context==4.3.0
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.7
more-itertools==9.1.0
MouseInfo==0.1.3
nest-asyncio==1.6.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==24.1
parso==0.8.4
Pillow==9.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
PyAudio==0.2.13
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.18.0
pyjokes==0.6.0
PyJWT==2.6.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.7.1
pywhatkit==5.4
pywin32==305
pyzmq==26.2.0
requests==2.28.2
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4
SpeechRecognition==3.10.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
twilio==7.16.5
typing_extensions==4.12.2
urllib3==1.26.15
wcwidth==0.2.13
Werkzeug==2.2.3
wikipedia==1.4.0
winshell==0.6
winspeech==1.0.1
wolframalpha==5.0.0
xmltodict==0.13.0
